+++ dart/lib/src/utils/add_tracing_headers_to_http_request.dart
@@ -1,0 +1,29 @@
+import 'package:meta/meta.dart';
+
+import '../../sentry.dart';
+
+@internal
+void addTracingHeadersToHttpHeader(Map<String, dynamic> headers, Hub hub,
+    {ISentrySpan? span}) {
+  if (span != null) {
+    addSentryTraceHeaderFromSpan(span, headers);
+    addW3CHeaderFromSpan(span, headers);
+    addBaggageHeaderFromSpan(
+      span,
+      headers,
+      logger: hub.options.logger,
+    );
+  } else {
+    final scope = hub.scope;
+    final propagationContext = scope.propagationContext;
+
+    final traceHeader = propagationContext.toSentryTrace();
+    addSentryTraceHeader(traceHeader, headers);
+    addW3CHeaderFromSentryTrace(traceHeader, headers);
+
+    final baggageHeader = propagationContext.toBaggageHeader();
+    if (baggageHeader != null) {
+      addBaggageHeader(baggageHeader, headers, logger: hub.options.logger);
+    }
+  }
+}