+++ dart/test/sentry_test.dart
@@ -4,6 +4,7 @@ import 'dart:isolate';
 import 'package:sentry/sentry.dart';
 import 'package:sentry/src/dart_exception_type_identifier.dart';
 import 'package:sentry/src/event_processor/deduplication_event_processor.dart';
+import 'package:sentry/src/feature_flags_integration.dart';
 import 'package:test/test.dart';
 
 import 'mocks.dart';
@@ -268,6 +269,12 @@ void main() {
       );
       expect(
         optionsReference.integrations
+            .whereType<FeatureFlagsIntegration>()
+            .length,
+        1,
+      );
+      expect(
+        optionsReference.integrations
             .whereType<IsolateErrorIntegration>()
             .length,
         1,
@@ -293,6 +300,41 @@ void main() {
         },
       );
     }, onPlatform: {'vm': Skip()});
+
+    test('should add feature flag FeatureFlagsIntegration', () async {
+      await Sentry.init(
+        options: defaultTestOptions(),
+        (options) => options.dsn = fakeDsn,
+      );
+
+      await Sentry.addFeatureFlag('foo', true);
+
+      expect(
+        Sentry.currentHub.scope.contexts[SentryFeatureFlags.type]?.values.first
+            .name,
+        equals('foo'),
+      );
+      expect(
+        Sentry.currentHub.scope.contexts[SentryFeatureFlags.type]?.values.first
+            .value,
+        equals(true),
+      );
+    });
+
+    test('addFeatureFlag should ignore non-boolean values', () async {
+      await Sentry.init(
+        options: defaultTestOptions(),
+        (options) => options.dsn = fakeDsn,
+      );
+
+      await Sentry.addFeatureFlag('foo1', 'some string');
+      await Sentry.addFeatureFlag('foo2', 123);
+      await Sentry.addFeatureFlag('foo3', 1.23);
+
+      final featureFlagsContext =
+          Sentry.currentHub.scope.contexts[SentryFeatureFlags.type];
+      expect(featureFlagsContext, isNull);
+    });
 
     test('should close integrations', () async {
       final integration = MockIntegration();