+++ dio/lib/src/tracing_client_adapter.dart
@@ -1,4 +1,4 @@
-// ignore_for_file: strict_raw_type
+// ignore_for_file: strict_raw_type, invalid_use_of_internal_member
 
 import 'dart:typed_data';
 
@@ -6,13 +6,22 @@ import 'package:dio/dio.dart';
 import 'package:sentry/sentry.dart';
 
 /// A [Dio](https://pub.dev/packages/dio)-package compatible HTTP client adapter
-/// which adds support to Sentry Performance feature.
+/// which adds support to Sentry Performance feature. If tracing is disabled
+/// generated spans will be no-op. This client also handles adding the
+/// Sentry trace headers to the HTTP request header.
 /// https://develop.sentry.dev/sdk/performance
 class TracingClientAdapter implements HttpClientAdapter {
   // ignore: public_member_api_docs
+  static const String integrationName = 'HTTPNetworkTracing';
+
+  // ignore: public_member_api_docs
   TracingClientAdapter({required HttpClientAdapter client, Hub? hub})
       : _hub = hub ?? HubAdapter(),
-        _client = client;
+        _client = client {
+    if (_hub.options.isTracingEnabled()) {
+      _hub.options.sdk.addIntegration(integrationName);
+    }
+  }
 
   final HttpClientAdapter _client;
   final Hub _hub;
@@ -23,7 +32,6 @@ class TracingClientAdapter implements HttpClientAdapter {
     Stream<Uint8List>? requestStream,
     Future? cancelFuture,
   ) async {
-    // ignore: invalid_use_of_internal_member
     final urlDetails = HttpSanitizer.sanitizeUrl(options.uri.toString());
 
     var description = options.method;
@@ -38,61 +46,29 @@ class TracingClientAdapter implements HttpClientAdapter {
       description: description,
     );
 
-    // ignore: invalid_use_of_internal_member
-    span?.origin = SentryTraceOrigins.autoHttpDioHttpClientAdapter;
-
-    // if the span is NoOp, we don't want to attach headers
     if (span is NoOpSentrySpan) {
       span = null;
     }
 
+    // Regardless whether tracing is enabled or not, we always want to attach
+    // Sentry trace headers (tracing without performance).
+    if (containsTargetOrMatchesRegExp(
+      _hub.options.tracePropagationTargets,
+      options.uri.toString(),
+    )) {
+      addTracingHeadersToHttpHeader(options.headers, _hub, span: span);
+    }
+
+    span?.origin = SentryTraceOrigins.autoHttpDioHttpClientAdapter;
     span?.setData('http.request.method', options.method);
     urlDetails?.applyToSpan(span);
 
     ResponseBody? response;
     try {
-      if (containsTargetOrMatchesRegExp(
-        // ignore: invalid_use_of_internal_member
-        _hub.options.tracePropagationTargets,
-        options.uri.toString(),
-      )) {
-        if (span != null) {
-          addSentryTraceHeaderFromSpan(span, options.headers);
-          addW3CHeaderFromSpan(span, request.headers);
-          addBaggageHeaderFromSpan(
-            span,
-            options.headers,
-            // ignore: invalid_use_of_internal_member
-            logger: _hub.options.logger,
-          );
-        } else {
-          // ignore: invalid_use_of_internal_member
-          final scope = _hub.scope;
-          // ignore: invalid_use_of_internal_member
-          final propagationContext = scope.propagationContext;
-
-          final traceHeader = propagationContext.toSentryTrace();
-          addSentryTraceHeader(traceHeader, options.headers);
-          addW3CHeaderFromSentryTrace(traceHeader, options.headers);
-
-          final baggage = propagationContext.baggage;
-          if (baggage != null) {
-            final baggageHeader = SentryBaggageHeader.fromBaggage(baggage);
-            addBaggageHeader(
-              baggageHeader,
-              options.headers,
-              // ignore: invalid_use_of_internal_member
-              logger: _hub.options.logger,
-            );
-          }
-        }
-      }
-
       response = await _client.fetch(options, requestStream, cancelFuture);
       span?.status = SpanStatus.fromHttpStatusCode(response.statusCode);
       span?.setData('http.response.status_code', response.statusCode);
       final contentLengthHeader =
-          // ignore: invalid_use_of_internal_member
           HttpHeaderUtils.getContentLength(response.headers);
       if (contentLengthHeader != null) {
         span?.setData('http.response_content_length', contentLengthHeader);